Use $projectDir in genPolicyStubs.dst
Currently, the genPolicyStubs.dst is defined as "src/main/java/sapphire/policy/stubs/"
which is a path relative to the current directory, i.e. the directory where the gradle
command is invoked. If a user runs `gradlew genPolicyStubs` outside of sapphire-core
project, the stub files will be put outside of sapphire-core.

This PR changes dst to "$projectDir/src/main/java/sapphire/policy/stubs/" so that
generated stub files always live in the src directory of sapphire-core project.