2018-8-20 Update to sapphire_replicate to update hostname and replicate to call default server policy
1. Update hostname for policy stub and its host in the next stub in the client.
This is because otherwise the hostname associated with the client stub will not be passed to the next client.
2. Update to use the current server policy to call sapphire_pin instead of using the method in replica. This is because replica might be already placed at a different location.

Current test will fail when it tries to get the todoList of partition key "2" because Kernel server for "2" does not have object reference for atLeastOneRpc but only for DHT2.

TODO: Update objectmanager in kernelserver to always point to the newly chained first server policy (last in client policy order).

Currently, only a single object is pinned (DHT2.sapphire_pin()) to the new server. Therefore, the second policy in the client side (atLeastOneRpc) will not be able to find server policy when it calls the second Kernel server. This is because atLeastOneRpc will call server policy for atLeastOneRpc in the server but in kernel server 2, only the server policy for DHT2 was registered to the object manager.

To achieve this, it is probably needed to update object manager to only have the first server policy as other server policy will not be called anyway. Object manager needs to be updated to point to newly added server policy (and remove the previous reference).

For DHT2, it calls sapphire_pin() after replcate(). Object movement is performed at sapphire_pin(). It means object movement stage needs to know the head pointer of the server policy, and copies the head pointer only.