2018-9-15 (in-progress) fix to access via method
1. Direct access to the variable (nextServerPolicy, previousServerPolicy) seems resulting in null even though they exist. Therefore, changing to accessing them via method.

2. Updated explicitReplicationPolicy to propagate onRPC call to all child servers to test ConsensusPolicy scenario.

TODO: #1 results in returning incorrect object when policy propagates update to all child servers via onRPC. This is because the call was delivered to the unintended servers as well. Therefore, we need to support direct call from group policy to servers. We may also need to support direct call from server to other servers.